added new "prepare-release" profile
Added a profiles section with a new profile called "prepare-release".
Moved the generation of the sources and javadoc JAR files and GPG
signing to this profile.  Only want to execute these when we're doing a
real release of this bundle, not when doing local development.  To use
this profile, type "mvn install -Pprepare-release".