upgrade to latest version of maven-gpg-plugin
Went from 1.4 to 1.5.  Note for Windows users - you need to specify the
passphrase on the command line, e.g., -Dgpg.passphrase=thephrase.  If
you left this out in 1.4, the plug-in would prompt you for the
passphrase.  In 1.5, the plug-in will sometimes hang if the passphrase
is not specified on the command line.  Don't know if this is specific to
Windows - did not test other platforms.